// src/app/model/pago.ts
export class Pago {
  id: number = 0;
  createdBy: number = 0;
  modifiedBy: number = 0;
  createDate: Date = new Date();
  modificationDate: Date = new Date();
  numeroPago: number = 0;
  fecha: Date = new Date();
  mensualidad: number = 0;
  mensualidadEnganche: number = 0;
  totalMensualidad: number = 0;
  saldoCapital: number = 0;
  fechaPago: Date = new Date();
  montoPagado: number = 0;
}